+++ src/common/components/MovieSearchPrompt.vue
@@ -60,7 +60,7 @@
 <script setup lang="ts">
 import { ref, computed, watch } from "vue";
 import { MovieSearchIndex } from "@/common/types/models";
-import { useSearch } from "@/data/useTMDB";
+import { useSearch } from "@/service/useTMDB";
 
 const { defaultList, defaultListTitle } = defineProps<{
   defaultList: MovieSearchIndex[];
