+++ src/features/clubs/views/ClubsView.vue
@@ -24,8 +24,8 @@ import { computed, reactive, watch } from "vue";
 import { useStore } from "vuex";
 
 import clubSvg from "@/assets/images/menu-images/club.svg";
-import { useUser } from "@/data/useUser";
-import { useClub } from "@/data/useClub";
+import { useUser } from "@/service/useUser";
+import { useClub } from "@/service/useClub";
 import { ClubsViewClub, CacheDataService } from "@/common/types/models";
 
 const store = useStore();
