+++ src/features/reviews/components/AddReviewPrompt.vue
@@ -17,8 +17,8 @@ import { computed } from "vue";
 import { MovieSearchIndex } from "@/common/types/models";
 import MovieSearchPrompt from "../../../common/components/MovieSearchPrompt.vue";
 import { useRoute } from "vue-router";
-import { useDeleteMovie, useWatchList } from "@/data/useWatchList";
-import { useAddReview } from "@/data/useReview";
+import { useDeleteMovie, useWatchList } from "@/service/useWatchList";
+import { useAddReview } from "@/service/useReview";
 
 const emit = defineEmits<{
   (e: "close"): void;
